26-1-5.1：更新依赖，插入KafkaListener的中文注释

+++ finalAssignmentBackend/src/main/java/com/tutict/finalassignmentbackend/kafkaListener/DriverInformationKafkaListener.java
@@ -14,6 +14,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 @Service
+// Kafka 监听器，处理消息
 public class DriverInformationKafkaListener {
 
     private static final Logger log = Logger.getLogger(DriverInformationKafkaListener.class.getName());
@@ -21,6 +22,7 @@ public class DriverInformationKafkaListener {
     private final DriverInformationService driverInformationService;
     private final ObjectMapper objectMapper;
 
+    // 构造器注入依赖
     @Autowired
     public DriverInformationKafkaListener(DriverInformationService driverInformationService,
                                           ObjectMapper objectMapper) {
@@ -28,20 +30,25 @@ public class DriverInformationKafkaListener {
         this.objectMapper = objectMapper;
     }
 
+    // 监听 Kafka 消息
     @KafkaListener(topics = "driver_information_create", groupId = "driverInformationGroup", concurrency = "3")
     public void onDriverInformationCreate(@Header(value = KafkaHeaders.RECEIVED_KEY, required = false) byte[] rawKey,
                                           @Payload String message) {
         log.log(Level.INFO, "Received Kafka message for DriverInformation create: {0}", message);
+        // 使用虚拟线程异步处理，避免阻塞监听线程
         Thread.ofVirtual().start(() -> processMessage(asKey(rawKey), message, "create"));
     }
 
+    // 监听 Kafka 消息
     @KafkaListener(topics = "driver_information_update", groupId = "driverInformationGroup", concurrency = "3")
     public void onDriverInformationUpdate(@Header(value = KafkaHeaders.RECEIVED_KEY, required = false) byte[] rawKey,
                                           @Payload String message) {
         log.log(Level.INFO, "Received Kafka message for DriverInformation update: {0}", message);
+        // 使用虚拟线程异步处理，避免阻塞监听线程
         Thread.ofVirtual().start(() -> processMessage(asKey(rawKey), message, "update"));
     }
 
+    // 统一处理消息并执行业务逻辑
     private void processMessage(String idempotencyKey, String message, String action) {
         if (isBlank(idempotencyKey)) {
             log.warning("Received driver information event without idempotency key, skipping");
@@ -69,7 +76,7 @@ public class DriverInformationKafkaListener {
                 return;
             }
             driverInformationService.markHistorySuccess(idempotencyKey,
-                    result.getDriverId() != null ? result.getDriverId().longValue() : null);
+                    result.getDriverId() != null ? result.getDriverId() : null);
         } catch (Exception ex) {
             driverInformationService.markHistoryFailure(idempotencyKey, ex.getMessage());
             log.log(Level.SEVERE,
@@ -79,6 +86,7 @@ public class DriverInformationKafkaListener {
         }
     }
 
+    // 反序列化消息体
     private DriverInformation deserializeMessage(String message) {
         try {
             return objectMapper.readValue(message, DriverInformation.class);
@@ -88,10 +96,12 @@ public class DriverInformationKafkaListener {
         }
     }
 
+    // 将 Kafka key 转为字符串
     private String asKey(byte[] rawKey) {
         return rawKey == null ? null : new String(rawKey);
     }
 
+    // 判空
     private boolean isBlank(String value) {
         return value == null || value.trim().isEmpty();
     }
